fix reward being int

+++ demonstration.py
@@ -60,7 +60,7 @@ class Demonstration(object):
             state_shape = self.states[0].shape
             S = f.create_dataset('S', (len(self), ) + state_shape, dtype='uint8', compression='lzf', data=np.array(self.states))
             A = f.create_dataset('A', (len(self), ), dtype='uint8', data=np.array(self.actions))
-            R = f.create_dataset('R', (len(self), ), dtype='int32', data=np.array(self.rewards))
+            R = f.create_dataset('R', (len(self), ), dtype='float32', data=np.array(self.rewards))
             terminal = f.create_dataset('terminal', (len(self), ), dtype='b', data=np.array(self.terminals))
             lives = f.create_dataset('lives', (len(self), ), dtype='uint8', data=np.array(self.lives))
             # emulator state
